+++ app/CMakeLists.txt
@@ -13,7 +13,7 @@ FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
 
 # Add display module only if enabled
-FILE(GLOB display_sources src/display/*.c src/images/*.c)
+FILE(GLOB display_sources src/display/*.c src/images/*.c src/ui/*.c src/ui/screens/*.c)
 target_sources_ifdef(CONFIG_HEALTHYPI_DISPLAY_ENABLED app PRIVATE ${display_sources})
 
 FILE(GLOB bt_sources src/bt/*.c)
@@ -21,4 +21,5 @@ target_sources_ifdef(CONFIG_BT app PRIVATE ${bt_sources})
 
 zephyr_include_directories(src)
 zephyr_include_directories(src/display)
+zephyr_include_directories(src/ui)
 zephyr_include_directories(src/bt)
