+++ make_disp.sh
@@ -1,1 +1,1 @@
-west build -p always -b healthypi5_rp2040 app -DEXTRA_CONF_FILE=overlay-display.conf -DEXTRA_DTC_OVERLAY_FILE=healthypi5_rp2040_display.overlay
+west build -p auto -b healthypi5_rp2040 app -DEXTRA_CONF_FILE=overlay-display.conf -DEXTRA_DTC_OVERLAY_FILE=healthypi5_rp2040_display.overlay
